+++ src/Appstore/Bundle/MedicineBundle/Controller/MedicinePurchaseReturnController.php
@@ -214,7 +214,7 @@ class MedicinePurchaseReturnController extends Controller
 
 		$em = $this->getDoctrine()->getManager();
 		$discountType = $request->request->get('discountType');
-		$discountCal = (int)$request->request->get('discount');
+		$discountCal = (float)$request->request->get('discount');
 		$invoice = $request->request->get('purchase');
 		$entity = $em->getRepository('MedicineBundle:MedicinePurchaseReturn')->find($invoice);
 		$subTotal = $entity->getSubTotal();
